% basename
\name{}

% topic(s) documented
\alias{}

% title of the topic(s)
\title{}

% short description
\description{}

% function synopsis
\usage{
   
}

% function arguments
\arguments{
   \item{}{}
}

% detailed description of the functionality
\details{}

% description of the return value(s)
\value{}

% literature references
%\references{}

% special note(s)
%\note{}

% author of this file
\author{Daniel Haase.}

% links to related R objects
%\seealso{}

% code examples
\examples{
   
}

% R documentation keywords
%\keyword{}
