\name{deg.rad}

\alias{deg2rad}
\alias{rad2deg}

\title{Conversion between radian and degrees}

\description{Convert radian to degrees and vice versa.}

\usage{
   deg2rad(deg)
   rad2deg(rad)   
}

\arguments{
   \item{deg}{Numeric object (e.g. vector, matrix, array) giving the angles in degrees.}
   \item{rad}{Numeric object giving the angles in radian.}
}

\value{A numerical object of the same size as the argument, giving the angles in the converted unit.}
