% basename
\name{function-along}

% topic(s) documented
\alias{along}

% title of the topic(s)
\title{Sequence Generation}

% short description
\description{Alias for \code{seq_along}.}

% function synopsis
\usage{
   along(along.with)
}

% function arguments
\arguments{
   \item{along.with}{Argument to be passed to \code{seq_along}.}
}

% detailed description of the functionality
%\details{}

% description of the return value(s)
\value{See \code{\link{seq_along}}.}

% literature references
%\references{}

% special note(s)
%\note{}

% author of this file
\author{Daniel Haase.}

% links to related R objects
\seealso{\code{\link{seq_along}}.}

% code examples
\examples{
   along(character(0))
   along(10)
   along(c(7, 19, 201))
}

% R documentation keywords
%\keyword{}
