\name{Test Objects}

\alias{M}

\title{Example objects for testing purposes}

\description{Example R objects for testing and debugging purposes.}

\usage{M}

\details{
   \describe{
      \item{\code{M}}{\eqn{4\times4}{4x4} matrix with values \eqn{1} to \eqn{16}.}
   }
}

\examples{
   M
   apply(M, 1, mean)
}
