\name{Undocumented}

\alias{bytes.toHumanReadable}
\alias{colorhash}
\alias{colorize}
\alias{desub}
\alias{dirname<-}
\alias{entropy}
\alias{equal}
\alias{file.extension<-}
\alias{fill}
\alias{gaussians2d}
\alias{init}
\alias{install.suggestedPackages}
\alias{interval.transform}
\alias{is.directory}
\alias{kurtosis}
\alias{match.args}
\alias{minRequiredVersion}
\alias{mvdnorm}
\alias{objects.sizes}
\alias{out}
\alias{pal}
\alias{palettes}
\alias{paste0}
\alias{pm}
\alias{pr}
\alias{progressBar}
\alias{readImage}
\alias{reload.package}
\alias{scurve}
\alias{sec.dhms}
\alias{st}
\alias{step}
\alias{time.ymdhms}
\alias{triangle.area}
\alias{writeImage}

\title{Undocumented Code Objects of Package \code{dh}}

\description{These objects are undocumented.}
